+++ DTOs/ModifyCompanyDTO.cs
@@ -3,7 +3,7 @@
     public class ModifyCompanyDTO
     {
         public string CompanyName { get; set; }
-        public int CompanyAddressId { get; set; }
+        public ulong CompanyAddressId { get; set; }
         public string Website { get; set; }
         public string Password { get; set; }
         public string Desciption { get; set; }
