+++ DTOs/ModifyUserDTO.cs
@@ -6,7 +6,7 @@
         public string LastName { get; set; }
         public DateTime BirthdayDate { get; set; }
         public string ContactNumber { get; set; }
-        public int UserAddressId { get; set; }
+        public ulong UserAddressId { get; set; }
         public string Gender { get; set; }
         public float Experience { get; set; }
         public string Password { get; set; }
